+++ mlflow/tracing/export/inference_table.py
@@ -27,9 +27,9 @@ def pop_trace(request_id: str) -> Optional[Dict[str, Any]]:
 
 
 # For Inference Table, we use special TTLCache to store the finished traces
-# so that they can be retrieved by the Databricks model serving. The values
-# in the buffer is not Trace dataclass but a dictionary with the schema
-# used in the Databricks.
+# so that they can be retrieved by Databricks model serving. The values
+# in the buffer are not Trace dataclass, but rather a dictionary with the schema
+# that is used within Databricks model serving.
 def _initialize_trace_buffer():  # Define as a function for testing purposes
     return TTLCache(
         maxsize=MLFLOW_TRACE_BUFFER_MAX_SIZE.get(),
